Extract grab_gui_element into new irr_gui_ptr.h
